# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger documentatie van de Overhoorder API (Node.js versie)",
    "version": "1.1.0",
    "title": "Overhoorder API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
...
  "paths": {
...
    "/admin/create-license": {
      "post": {
        "tags": ["admin"],
        "summary": "Licentie aanmaken",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "licentie_type": { "type": "string" },
                "docent_id": { "type": "integer" }
              },
              "required": ["licentie_type"]
            }
          }
        ],
        "responses": { "200": { "description": "Licentie aangemaakt" } }
      }
    },
...
    "/admin/licenses": {
      "get": {
        "tags": ["admin"],
        "summary": "Licenties ophalen",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Licenties" } }
      }
    },
...
  },
...
}